refactor: remove v1 implementation and use v2 (with PKCE) as the default
