+++ test/SgidClient.spec.ts
@@ -1,27 +1,32 @@
 import { readFileSync } from 'fs'
 
-import SgidClient, { DEFAULT_SCOPE } from '../src'
+import SgidClient from '../src'
+import {
+  DEFAULT_SCOPE,
+  DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+} from '../src/constants'
+import { codeVerifierAndChallengePattern } from '../src/util'
 
 import {
   MOCK_ACCESS_TOKEN,
-  MOCK_API_VERSION,
   MOCK_AUTH_CODE,
-  MOCK_AUTH_ENDPOINT,
+  MOCK_AUTH_ENDPOINT_V2,
   MOCK_CLIENT_ID,
   MOCK_CLIENT_PRIVATE_KEY,
   MOCK_CLIENT_SECRET,
+  MOCK_CODE_VERIFIER,
   MOCK_HOSTNAME,
   MOCK_REDIRECT_URI,
   MOCK_SUB,
   MOCK_USERINFO_PLAINTEXT,
 } from './mocks/constants'
 import {
-  tokenHandlerNoSub,
-  tokenHandlerNoToken,
-  userInfoHandlerMalformedData,
-  userInfoHandlerMalformedKey,
-  userInfoHandlerNoData,
-  userInfoHandlerNoKey,
+  tokenHandlerNoSubV2,
+  tokenHandlerNoTokenV2,
+  userInfoHandlerMalformedDataV2,
+  userInfoHandlerMalformedKeyV2,
+  userInfoHandlerNoDataV2,
+  userInfoHandlerNoKeyV2,
 } from './mocks/handlers'
 import { server } from './mocks/server'
 
@@ -35,9 +40,8 @@ describe('SgidClient', () => {
       clientId: MOCK_CLIENT_ID,
       clientSecret: MOCK_CLIENT_SECRET,
       privateKey: MOCK_CLIENT_PRIVATE_KEY,
-      redirectUri: MOCK_REDIRECT_URI,
+      redirectUris: [MOCK_REDIRECT_URI],
       hostname: MOCK_HOSTNAME,
-      apiVersion: MOCK_API_VERSION,
     })
   })
 
@@ -55,26 +59,29 @@ describe('SgidClient', () => {
         clientId: MOCK_CLIENT_ID,
         clientSecret: MOCK_CLIENT_SECRET,
         privateKey: pkcs8Key,
-        redirectUri: MOCK_REDIRECT_URI,
+        redirectUris: [MOCK_REDIRECT_URI],
         hostname: MOCK_HOSTNAME,
-        apiVersion: MOCK_API_VERSION,
       })
 
       expect(pkcs8Client).toBeDefined()
     })
   })
 
-  describe('authorizationUrl', () => {
-    it('should generate authorisation URL correctly when only state is provided', () => {
+  describe('authorizationUrl (v2 with PKCE)', () => {
+    it('should generate authorisation URL correctly when state and codeChallenge is provided', () => {
       const mockState = 'mockState'
+      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl(mockState)
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        codeChallenge: mockCodeChallenge,
+      })
 
       const actual = new URL(url)
       // Count number of search params
       let actualNumSearchParams = 0
       actual.searchParams.forEach(() => actualNumSearchParams++)
-      const expected = new URL(MOCK_AUTH_ENDPOINT)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
       expect(actual.host).toBe(expected.host)
       expect(actual.pathname).toBe(expected.pathname)
       expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
@@ -85,21 +92,65 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      // Client ID, scope, response_type, redirect_uri, nonce, state
-      expect(actualNumSearchParams).toBe(6)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+      )
+      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(8)
     })
 
-    it('should generate authorisation URL correctly when scope is provided as a string', () => {
+    it('should generate authorisation URL correctly when state, codeChallenge, and codeChallengeMethod is provided', () => {
       const mockState = 'mockState'
-      const mockScope = 'mockScope'
+      const mockCodeChallenge = 'mockCodeChallenge'
+      const mockCodeChallengeMethod = 'plain'
 
-      const { url, nonce } = client.authorizationUrl(mockState, mockScope)
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        codeChallenge: mockCodeChallenge,
+        codeChallengeMethod: mockCodeChallengeMethod,
+      })
 
       const actual = new URL(url)
       // Count number of search params
       let actualNumSearchParams = 0
       actual.searchParams.forEach(() => actualNumSearchParams++)
-      const expected = new URL(MOCK_AUTH_ENDPOINT)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
+      expect(actual.host).toBe(expected.host)
+      expect(actual.pathname).toBe(expected.pathname)
+      expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
+      expect(actual.searchParams.get('scope')).toBe(DEFAULT_SCOPE)
+      expect(actual.searchParams.get('response_type')).toBe(
+        DEFAULT_RESPONSE_TYPE,
+      )
+      expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
+      expect(actual.searchParams.get('nonce')).toBe(nonce)
+      expect(actual.searchParams.get('state')).toBe(mockState)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      // Important check is here
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        mockCodeChallengeMethod,
+      )
+      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(8)
+    })
+
+    it('should generate authorisation URL correctly when state, codeChallenge, and scope is provided as a string', () => {
+      const mockState = 'mockState'
+      const mockScope = 'mockScope'
+      const mockCodeChallenge = 'mockCodeChallenge'
+
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        scope: mockScope,
+        codeChallenge: mockCodeChallenge,
+      })
+
+      const actual = new URL(url)
+      // Count number of search params
+      let actualNumSearchParams = 0
+      actual.searchParams.forEach(() => actualNumSearchParams++)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
       expect(actual.host).toBe(expected.host)
       expect(actual.pathname).toBe(expected.pathname)
       expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
@@ -111,21 +162,30 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      // Client ID, scope, response_type, redirect_uri, nonce, state
-      expect(actualNumSearchParams).toBe(6)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+      )
+      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(8)
     })
 
-    it('should generate authorisation URL correctly when scope is provided as a string array', () => {
+    it('should generate authorisation URL correctly when state, codeChallenge, and scope is provided as a string array', () => {
       const mockState = 'mockState'
       const mockScopes = ['mockScope1', 'mockScope2', 'mockScope3']
+      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl(mockState, mockScopes)
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        scope: mockScopes,
+        codeChallenge: mockCodeChallenge,
+      })
 
       const actual = new URL(url)
       // Count number of search params
       let actualNumSearchParams = 0
       actual.searchParams.forEach(() => actualNumSearchParams++)
-      const expected = new URL(MOCK_AUTH_ENDPOINT)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
       expect(actual.host).toBe(expected.host)
       expect(actual.pathname).toBe(expected.pathname)
       expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
@@ -137,25 +197,30 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      // Client ID, scope, response_type, redirect_uri, nonce, state
-      expect(actualNumSearchParams).toBe(6)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+      )
+      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(8)
     })
 
-    it('should generate authorisation URL correctly when nonce is specified', () => {
+    it('should generate authorisation URL correctly when state, codeChallenge, and nonce is specified', () => {
       const mockState = 'mockState'
       const mockNonce = 'mockNonce'
+      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl(
-        mockState,
-        undefined,
-        mockNonce,
-      )
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        nonce: mockNonce,
+        codeChallenge: mockCodeChallenge,
+      })
 
       const actual = new URL(url)
       // Count number of search params
       let actualNumSearchParams = 0
       actual.searchParams.forEach(() => actualNumSearchParams++)
-      const expected = new URL(MOCK_AUTH_ENDPOINT)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
       expect(actual.host).toBe(expected.host)
       expect(actual.pathname).toBe(expected.pathname)
       expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
@@ -168,20 +233,29 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('nonce')).toBe(mockNonce)
       expect(nonce).toBe(mockNonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      // Client ID, scope, response_type, redirect_uri, nonce, state
-      expect(actualNumSearchParams).toBe(6)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+      )
+      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(8)
     })
 
-    it('should generate authorisation URL correctly when nonce is null', () => {
+    it('should generate authorisation URL correctly when state and codeChallenge is provided and nonce is null', () => {
       const mockState = 'mockState'
+      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl(mockState, undefined, null)
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        nonce: null,
+        codeChallenge: mockCodeChallenge,
+      })
 
       const actual = new URL(url)
       // Count number of search params
       let actualNumSearchParams = 0
       actual.searchParams.forEach(() => actualNumSearchParams++)
-      const expected = new URL(MOCK_AUTH_ENDPOINT)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
       expect(actual.host).toBe(expected.host)
       expect(actual.pathname).toBe(expected.pathname)
       expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
@@ -194,26 +268,30 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('nonce')).toBeNull()
       expect(nonce).toBeUndefined()
       expect(actual.searchParams.get('state')).toBe(mockState)
-      // Client ID, scope, response_type, redirect_uri, state
-      expect(actualNumSearchParams).toBe(5)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+      )
+      // Client ID, scope, response_type, redirect_uri, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(7)
     })
 
-    it('should generate authorisation URL correctly when redirectUri is provided', () => {
+    it('should generate authorisation URL correctly when state, codeChallenge, and redirectUri is provided', () => {
       const mockState = 'mockState'
       const mockRedirectUri = 'https://mockRedirectUri.com'
+      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl(
-        mockState,
-        undefined,
-        undefined,
-        mockRedirectUri,
-      )
+      const { url, nonce } = client.authorizationUrl({
+        state: mockState,
+        redirectUri: mockRedirectUri,
+        codeChallenge: mockCodeChallenge,
+      })
 
       const actual = new URL(url)
       // Count number of search params
       let actualNumSearchParams = 0
       actual.searchParams.forEach(() => actualNumSearchParams++)
-      const expected = new URL(MOCK_AUTH_ENDPOINT)
+      const expected = new URL(MOCK_AUTH_ENDPOINT_V2)
       expect(actual.host).toBe(expected.host)
       expect(actual.pathname).toBe(expected.pathname)
       expect(actual.searchParams.get('client_id')).toBe(MOCK_CLIENT_ID)
@@ -225,49 +303,71 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(mockRedirectUri)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      // Client ID, scope, response_type, redirect_uri, nonce, state
-      expect(actualNumSearchParams).toBe(6)
+      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
+      expect(actual.searchParams.get('code_challenge_method')).toBe(
+        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
+      )
+      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
+      expect(actualNumSearchParams).toBe(8)
     })
 
     it('should throw when no redirectUri is provided', () => {
       const mockState = 'mockState'
+      const mockCodeChallenge = 'mockCodeChallenge'
 
       const noRedirectUriClient = new SgidClient({
         clientId: MOCK_CLIENT_ID,
         clientSecret: MOCK_CLIENT_SECRET,
         privateKey: MOCK_CLIENT_PRIVATE_KEY,
         hostname: MOCK_HOSTNAME,
-        apiVersion: MOCK_API_VERSION,
       })
 
-      expect(() => noRedirectUriClient.authorizationUrl(mockState)).toThrow(
-        'No redirect URI registered with this client',
-      )
+      expect(() =>
+        noRedirectUriClient.authorizationUrl({
+          state: mockState,
+          codeChallenge: mockCodeChallenge,
+        }),
+      ).toThrow('No redirect URI registered with this client')
     })
   })
 
-  describe('callback', () => {
+  describe('callback (v2 with PKCE)', () => {
     it('should call token endpoint and return sub and accessToken', async () => {
-      const { sub, accessToken } = await client.callback(MOCK_AUTH_CODE)
+      const { sub, accessToken } = await client.callback({
+        code: MOCK_AUTH_CODE,
+        codeVerifier: MOCK_CODE_VERIFIER,
+      })
 
       expect(sub).toBe(MOCK_SUB)
       expect(accessToken).toBe(MOCK_ACCESS_TOKEN)
     })
 
-    it('should throw when no access token is returned', async () => {
-      server.use(tokenHandlerNoToken)
-
-      await expect(client.callback(MOCK_AUTH_CODE)).rejects.toThrow(
-        'Missing sub claim or access token',
+    it('should throw when no code verifier is provided', async () => {
+      await expect(client.callback({ code: MOCK_AUTH_CODE })).rejects.toThrow(
+        "Code verifier must be provided in 'callback' when using apiVersion 2",
       )
     })
 
-    it('should throw when sub is empty', async () => {
-      server.use(tokenHandlerNoSub)
+    it('should throw when no access token is returned', async () => {
+      server.use(tokenHandlerNoTokenV2)
 
-      await expect(client.callback(MOCK_AUTH_CODE)).rejects.toThrow(
-        'Missing sub claim or access token',
-      )
+      await expect(
+        client.callback({
+          code: MOCK_AUTH_CODE,
+          codeVerifier: MOCK_CODE_VERIFIER,
+        }),
+      ).rejects.toThrow('Missing sub claim or access token')
+    })
+
+    it('should throw when sub is empty', async () => {
+      server.use(tokenHandlerNoSubV2)
+
+      await expect(
+        client.callback({
+          code: MOCK_AUTH_CODE,
+          codeVerifier: MOCK_CODE_VERIFIER,
+        }),
+      ).rejects.toThrow('Missing sub claim or access token')
     })
   })
 
@@ -280,7 +380,7 @@ describe('SgidClient', () => {
     })
 
     it('should return empty data object when no key is returned', async () => {
-      server.use(userInfoHandlerNoKey)
+      server.use(userInfoHandlerNoKeyV2)
 
       const { sub, data } = await client.userinfo(MOCK_ACCESS_TOKEN)
 
@@ -289,7 +389,7 @@ describe('SgidClient', () => {
     })
 
     it('should return empty data object when no data is returned', async () => {
-      server.use(userInfoHandlerNoData)
+      server.use(userInfoHandlerNoDataV2)
 
       const { sub, data } = await client.userinfo(MOCK_ACCESS_TOKEN)
 
@@ -302,9 +402,8 @@ describe('SgidClient', () => {
         clientId: MOCK_CLIENT_ID,
         clientSecret: MOCK_CLIENT_SECRET,
         privateKey: 'malformed',
-        redirectUri: MOCK_REDIRECT_URI,
+        redirectUris: [MOCK_REDIRECT_URI],
         hostname: MOCK_HOSTNAME,
-        apiVersion: MOCK_API_VERSION,
       })
 
       await expect(
@@ -313,7 +412,7 @@ describe('SgidClient', () => {
     })
 
     it('should throw when encrypted key is malformed', async () => {
-      server.use(userInfoHandlerMalformedKey)
+      server.use(userInfoHandlerMalformedKeyV2)
 
       await expect(client.userinfo(MOCK_ACCESS_TOKEN)).rejects.toThrow(
         'Unable to decrypt or import payload key',
@@ -321,11 +420,60 @@ describe('SgidClient', () => {
     })
 
     it('should throw when encrypted data is malformed', async () => {
-      server.use(userInfoHandlerMalformedData)
+      server.use(userInfoHandlerMalformedDataV2)
 
       await expect(client.userinfo(MOCK_ACCESS_TOKEN)).rejects.toThrow(
         'Unable to decrypt payload',
       )
+    })
+  })
+
+  describe('generatePkcePair', () => {
+    it('can generate a PKCE pair', () => {
+      const pkcePair = SgidClient.generatePkcePair()
+      expect(pkcePair).toBeDefined()
+    })
+  })
+
+  describe('generateCodeVerifier', () => {
+    it('should generate a code verifier of length 43 when no length is provided', () => {
+      const codeVerifier = SgidClient.generateCodeVerifier()
+
+      expect(codeVerifier.length).toBe(43)
+      expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
+    })
+
+    it('should generate a code verifier of specified length when length between 43 (inclusive) and 128 (inclusive) is provided', () => {
+      for (let length = 43; length <= 128; length++) {
+        const codeVerifier = SgidClient.generateCodeVerifier(length)
+        expect(codeVerifier.length).toBe(length)
+        expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
+      }
+    })
+
+    it('should throw an error when a length < 43 or length > 128 is provided', () => {
+      for (const length of [-1, 0, 42, 129, 138, 999]) {
+        expect(() => SgidClient.generateCodeVerifier(length)).toThrowError(
+          `The code verifier should have a minimum length of 43 and a maximum length of 128. Length of ${length} was provided`,
+        )
+      }
+    })
+  })
+
+  describe('generateCodeChallenge', () => {
+    it('should match the specified pattern', () => {
+      expect(SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)).toMatch(
+        codeVerifierAndChallengePattern,
+      )
+    })
+
+    it('should be deterministic (return the same code challenge given the same code verifier)', () => {
+      const firstCodeChallenge =
+        SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)
+      const secondCodeChallenge =
+        SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)
+
+      expect(firstCodeChallenge).toBe(secondCodeChallenge)
     })
   })
 })
